+++ app/Http/Controllers/Backend/SubCategoryController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\backend\CreateSubCategoryRequest;
 use App\Http\Requests\Backend\UpdateSubCategoryRequest;
 use App\Models\Category;
+use App\Models\ChildCategory;
 use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Redirect;
@@ -87,6 +88,10 @@ class SubCategoryController extends Controller
   public function destroy(string $id)
   {
     $subCategory = SubCategory::findOrFail($id);
+    $childCategory = ChildCategory::where('sub_category_id', $subCategory->id)->count();
+    if ($childCategory > 0) {
+      return response(['status' => 'error', 'message' => 'This Sub-Category contain sub items for delete this you need to delete the sub items first!']);
+    }
     $subCategory->delete();
 
     return response(['status' => 'success', 'message' => 'Deleted Successfully!']);
